add exit_loop_

+++ src/userspace/console_user_server/console_user_server/include/receiver.hpp
@@ -5,17 +5,19 @@
 
 class receiver final {
 public:
-  void start(void) {
-    try {
-      const char* path = constants::get_console_user_socket_file_path();
-      unlink(path);
-      server_ = std::make_unique<local_datagram_server>(path);
-      chmod(path, 0600);
+  receiver(void) : exit_loop_(false) {}
 
-      thread_ = std::thread([this] { this->worker(); });
-    } catch (...) {
-      std::cout << "default exception";
+  void start(void) {
+    if (server_) {
+      return;
     }
+
+    const char* path = constants::get_console_user_socket_file_path();
+    unlink(path);
+    server_ = std::make_unique<local_datagram_server>(path);
+    chmod(path, 0600);
+
+    thread_ = std::thread([this] { this->worker(); });
   }
 
   void stop(void) {
@@ -26,17 +28,17 @@ public:
       return;
     }
 
-    server_->get_socket().cancel();
+    exit_loop_ = true;
     thread_.join();
     server_.reset(nullptr);
   }
 
   void worker(void) {
-    for (;;) {
-      if (!server_) {
-        break;
-      }
+    if (!server_) {
+      return;
+    }
 
+    while (!exit_loop_) {
       boost::system::error_code ec;
       std::size_t n = server_->receive(boost::asio::buffer(buffer_), boost::posix_time::seconds(1), ec);
 
@@ -55,4 +57,5 @@ private:
   std::array<uint8_t, buffer_length> buffer_;
   std::unique_ptr<local_datagram_server> server_;
   std::thread thread_;
+  volatile bool exit_loop_;
 };
